+++ esp_cam_sensor/include/esp_cam_ctlr_spi.h
@@ -26,15 +26,29 @@ typedef enum esp_cam_ctlr_spi_cam_intf {
 } esp_cam_ctlr_spi_cam_intf_t;
 
 /**
+ * @brief SPI CAM data I/O mode
+ *
+ * @note When using ESP_CAM_CTLR_SPI_CAM_INTF_SPI interface, only 1-bit mode is supported.
+ *       2-bit mode is only available with ESP_CAM_CTLR_SPI_CAM_INTF_PARLIO interface.
+ */
+typedef enum esp_cam_ctlr_spi_cam_io_mode {
+    ESP_CAM_CTLR_SPI_CAM_IO_MODE_1BIT = 0,
+    ESP_CAM_CTLR_SPI_CAM_IO_MODE_2BIT
+} esp_cam_ctlr_spi_cam_io_mode_t;
+
+/**
  * @brief SPI CAM configuration
  */
 typedef struct esp_cam_ctlr_spi_config {
     esp_cam_ctlr_spi_cam_intf_t intf;                   /*!< SPI CAM interface type */
 
+    esp_cam_ctlr_spi_cam_io_mode_t io_mode;             /*!< SPI CAM data I/O mode */
+
     spi_host_device_t spi_port;                         /*!< SPI port, this will be ignored when intf is ESP_CAM_CTLR_SPI_CAM_INTF_PARLIO */
     gpio_num_t spi_cs_pin;                              /*!< SPI CS pin or parlio valid signal pin */
     gpio_num_t spi_sclk_pin;                            /*!< SPI SCLK pin or parlio clock signal pin */
     gpio_num_t spi_data0_io_pin;                        /*!< SPI data0 I/O pin or parlio data0 signal pin */
+    gpio_num_t spi_data1_io_pin;                        /*!< SPI data1 I/O pin or parlio data1 signal pin. Only required when io_mode is ESP_CAM_CTLR_SPI_CAM_IO_MODE_2BIT. Set to GPIO_NUM_NC for 1-bit mode. */
 
     gpio_num_t reset_pin;                               /*!< Reset pin */
     gpio_num_t pwdn_pin;                                /*!< Power down pin */